requirements.txt
pbr>=0.5.21,<1.0
SQLAlchemy>=0.7.8,<=0.7.99
eventlet>=0.13.0
extras
kombu>=2.4.8
Routes>=1.12.3
WebOb>=1.2.3,<1.3
PasteDeploy>=1.5.0
Paste
netaddr
httplib2
lxml>=2.3
passlib
python-keystoneclient>=0.3.2
python-swiftclient>=1.5
iso8601>=0.1.4
jsonschema>=1.3.0,!=1.4.0
pexpect
oslo.config>=1.2.0
MySQL-python
Babel>=0.9.6
